PRL-6359: Moved old pages to better directory
