+++ e2e/pages/citizen/createCase/C100/c100ScreeningSections/alternativeRoutesPage.ts
@@ -1,8 +1,8 @@
 import { Page } from "@playwright/test";
-import { Selectors } from "../../common/selectors";
-import { Helpers } from "../../common/helpers";
-import AccessibilityTestHelper from "../../common/accessibilityTestHelper";
-import { AlternativeRoutesContent } from "../../fixtures/citizen/createCase/C100/c100ScreeningSections/alternativeRoutesContent";
+import { Selectors } from "../../../../../common/selectors";
+import { Helpers } from "../../../../../common/helpers";
+import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper";
+import { AlternativeRoutesContent } from "../../../../../fixtures/citizen/createCase/C100/c100ScreeningSections/alternativeRoutesContent";
 
 interface AlternativeRoutesPageOptions {
   page: Page;
