+++ e2e/pages/citizen/createCase/C100/c100ScreeningSections/legalRepresentationApplicationPage.ts
@@ -1,8 +1,8 @@
 import { Page } from "@playwright/test";
-import AccessibilityTestHelper from "../../common/accessibilityTestHelper";
-import { Selectors } from "../../common/selectors";
-import { LegalRepresentationApplicationContent } from "../../fixtures/citizen/createCase/C100/c100ScreeningSections/legalRepresentationApplicationContent";
-import { Helpers } from "../../common/helpers";
+import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper";
+import { Selectors } from "../../../../../common/selectors";
+import { LegalRepresentationApplicationContent } from "../../../../../fixtures/citizen/createCase/C100/c100ScreeningSections/legalRepresentationApplicationContent";
+import { Helpers } from "../../../../../common/helpers";
 
 enum inputIDs {
   radioYes = "#sq_legalRepresentationApplication",
